added support for binary io to streams with compression if desired

+++ utils/create_voc_step0.cpp
@@ -1,10 +1,5 @@
-/**
- * Date:  2016
- * Author: Rafael Muñoz Salinas
- * Description: demo application of DBoW3
- * License: see the LICENSE.txt file
- */
 
+//First step of creating a vocabulary is extracting features from a set of images. We save them to a file for next step
 #include <iostream>
 #include <vector>
 
@@ -66,7 +61,7 @@ vector< cv::Mat  >  loadFeatures( std::vector<string> path_to_images,string desc
 
 
     cout << "Extracting   features..." << endl;
-    for(int i = 0; i < path_to_images.size(); ++i)
+    for(size_t i = 0; i < path_to_images.size(); ++i)
     {
         vector<cv::KeyPoint> keypoints;
         cv::Mat descriptors;
